[Feature]: codebase cleanup
