+++ app/src/schemas/address.ts
@@ -1,5 +1,16 @@
 import { z } from "zod";
 import { AddressTypeSchema } from "./enums";
+import { phoneSchema } from "./common";
+
+export const addressSchema = z.object({
+    address_type: AddressTypeSchema,
+    first_name: z.string().min(1, "First name is required"),
+    last_name: z.string().min(1, "Last name is required"),
+    address_1: z.string().min(1, "Address is required"),
+    address_2: z.string().optional(),
+    state: z.string().min(1, "State is required"),
+    phone: phoneSchema,
+});
 
 export const AddressSchema = z.object({
     id: z.number(),
