+++ app/src/styles.css
@@ -64,8 +64,6 @@
 
     --background-image-gradient-primary: var(--gradient-primary);
     --background-image-gradient-action: var(--gradient-action);
-    --background-image-gradient-accent: var(--gradient-accent);
-    --background-image-gradient-subtle: var(--gradient-subtle);
 
     --aspect-product: 4 / 5;
     --aspect-gallery: 4 / 5.9;
@@ -75,17 +73,29 @@
 
     --radius-1xl: 0.875rem;
 
-    --animate-pulse-glow: pulse-glow 2s ease-in-out infinite;
     --animate-bounce: bounce 1s infinite;
-    --animate-shimmer: shimmer 2s infinite;
     --animate-float: float 6s ease-in-out infinite;
     --animate-fade-in: fade-in 0.8s ease-out;
     --animate-slide-up: slide-up 0.8s ease-out;
     --animate-scale-in: scale-in 0.4s ease-out;
-    --animate-sparkle-burst: sparkle-burst 1.5s ease-in-out 0s infinite normal none running;
-    --animate-slide-fade-in: slide-up 0.5s ease-out 0.2s both;
-    --animate-wave: wave 1.5s ease-in-out infinite;
     --animate-progress-bar: progress-bar 2s ease-out forwards;
+    --animate-border-spin: border-spin 3s linear infinite;
+    --animate-icon-bounce: icon-bounce 1.8s ease-in-out infinite;
+
+    @keyframes border-spin {
+        to {
+            transform: rotate(360deg);
+        }
+    }
+
+    @keyframes icon-bounce {
+        0%, 100% {
+            transform: translateX(0);
+        }
+        50% {
+            transform: translateX(3px);
+        }
+    }
 
     @keyframes float {
         0%,
@@ -110,27 +120,6 @@
         }
     }
 
-    @keyframes pulse-glow {
-        0%,
-        100% {
-            box-shadow: 0 0 20px hsl(var(--accent) / 0.3);
-        }
-
-        50% {
-            box-shadow: 0 0 40px hsl(var(--accent) / 0.6);
-        }
-    }
-
-    @keyframes shimmer {
-        0% {
-            background-position: -200% center;
-        }
-
-        100% {
-            background-position: 200% center;
-        }
-    }
-
     @keyframes scale-in {
         0% {
             transform: scale(0.8);
@@ -139,35 +128,6 @@
 
         100% {
             transform: scale(1);
-            opacity: 1;
-        }
-    }
-
-    @keyframes sparkle-burst {
-        0% {
-            opacity: 0;
-            transform: scale(0);
-        }
-
-        50% {
-            opacity: 1;
-            transform: scale(1.2);
-        }
-
-        100% {
-            opacity: 0;
-            transform: scale(0);
-        }
-    }
-
-    @keyframes wave {
-        0%,
-        100% {
-            transform: translateY(0) scale(1);
-            opacity: 0.3;
-        }
-        50% {
-            transform: translateY(-10px) scale(1.2);
             opacity: 1;
         }
     }
@@ -207,8 +167,6 @@
 
     --gradient-action: linear-gradient(135deg, hsl(340, 82%, 52%) 0%, hsl(15, 90%, 60%) 100%);
     --gradient-primary: linear-gradient(135deg, hsl(219 100% 53%), hsl(38 92% 50%));
-    --gradient-accent:  linear-gradient(135deg, hsl(38 92% 50%), hsl(43 96% 56%));
-    --gradient-subtle:  linear-gradient(135deg, hsl(38 92% 50% / 0.10), hsl(219 100% 53% / 0.07));
 
     --accent: 38 92% 50%;
     --accent-foreground: 20 14% 4%;
