--- app/src/types/generate.ts
@@ -1,4 +1,0 @@
-type PaginatedResponse<T> = {
-    next_cursor?: string | null;
-    items: T[];
-};
