+++ app/src/utils/index.ts
@@ -1,31 +1,5 @@
-import type { Message } from "@/schemas";
-
-const buildUrl = (baseUrl: string, queryParams: Record<string, string | number | Date | undefined | null>): string => {
-    let url = baseUrl;
-    let firstQueryParam = true;
-
-    for (const key in queryParams) {
-        if (!queryParams[key]) continue;
-
-        if (Object.hasOwn(queryParams, key)) {
-            if (firstQueryParam) {
-                url += `?${key}=${queryParams[key]}`;
-                firstQueryParam = false;
-            } else {
-                url += `&${key}=${queryParams[key]}`;
-            }
-        }
-    }
-
-    return url;
-};
-
 const currency = (number: number): string => {
     return number?.toLocaleString("en-NG", { style: "currency", currency: "NGN", maximumFractionDigits: 0 });
-};
-
-const capitalize = (str: string): string => {
-    return str.charAt(0).toUpperCase() + str.slice(1);
 };
 
 const debounce = <T extends (...args: any[]) => void>(func: T, delay: number): ((...args: Parameters<T>) => void) => {
@@ -37,52 +11,6 @@ const debounce = <T extends (...args: any[]) => void>(func: T, delay: number): (
             func(...args);
         }, delay);
     };
-};
-
-const isEqual = (value: any, other: any): boolean => {
-    if (value === other) {
-        return true;
-    }
-
-    if (value == null || other == null) {
-        return value === other;
-    }
-
-    if (typeof value !== typeof other) {
-        return false;
-    }
-
-    if (Array.isArray(value) && Array.isArray(other)) {
-        if (value.length !== other.length) {
-            return false;
-        }
-        for (let i = 0; i < value.length; i++) {
-            if (!isEqual(value[i], other[i])) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    if (typeof value === "object" && typeof other === "object") {
-        const valueKeys = Object.keys(value);
-        const otherKeys = Object.keys(other);
-
-        if (valueKeys.length !== otherKeys.length) {
-            return false;
-        }
-
-        for (const key of valueKeys) {
-            if (!otherKeys.includes(key) || !isEqual(value[key], other[key])) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    return value === other;
 };
 
 const omit = <T extends object, K extends keyof T>(obj: T, keys: K[] | K): Omit<T, K> => {
@@ -137,31 +65,6 @@ const formatDate = (dateString?: string) => {
     }).format(date);
 };
 
-const handleError = (error: any): Message => {
-    return { message: (error as Error).message || "An error occurred", error: true };
-};
-
-export const getPercentageDiff = (original: number, calculated: number) => {
-    const diff = original - calculated;
-    const decrease = (diff / original) * 100;
-
-    return decrease.toFixed();
-};
-
-export const isObject = (input: any) => input instanceof Object;
-export const isArray = (input: any) => Array.isArray(input);
-export const isEmpty = (input: any) => {
-    return (
-        input === null ||
-        input === undefined ||
-        (isObject(input) && Object.keys(input).length === 0) ||
-        (isArray(input) && (input as any[]).length === 0) ||
-        (typeof input === "string" && input.trim().length === 0)
-    );
-};
-
-export const onlyUnique = (value: unknown, index: number, self: unknown[]) => self.indexOf(value) === index;
-
 export const getInitials = (name: string) => {
     return name
         ?.split(" ")
@@ -169,13 +72,6 @@ export const getInitials = (name: string) => {
         ?.join("")
         ?.toUpperCase()
         ?.slice(0, 2);
-};
-
-export const getAvatarColor = (name: string) => {
-    const colors = ["bg-blue-500", "bg-green-500", "bg-purple-500", "bg-pink-500", "bg-yellow-500", "bg-indigo-500"];
-    const index = name.length % colors.length;
-
-    return colors[index];
 };
 
 export const formatTime = (input: Date | string | undefined): string => {
@@ -222,4 +118,4 @@ export const initializeAppSession = () => {
     }
 };
 
-export { handleError, capitalize, currency, buildUrl, debounce, isEqual, omit, generateId, timeAgo, formatDate };
+export { currency, debounce, omit, generateId, timeAgo, formatDate };
